Reorganize some of the BehaviorTree-related code (#957)
* Reorganize some of the BehaviorTree-related code.

Eliminate the NavigateToPoseBehaviorTree class, merging it with the BehaviorTreeEngine.
Put clear_entirely service client with the others in nav2_util.
Move RecoveryNode to nav2_behavior_tree with the other BT nodes.

* Remove method signature that isn't implemented yet

* Add a missing dependency

* Address a couple linter issues
