+++ nav2_behavior_tree/include/nav2_behavior_tree/recovery_node.hpp
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NAV2_BT_NAVIGATOR__RECOVERY_NODE_HPP_
-#define NAV2_BT_NAVIGATOR__RECOVERY_NODE_HPP_
+#ifndef NAV2_BEHAVIOR_TREE__RECOVERY_NODE_HPP_
+#define NAV2_BEHAVIOR_TREE__RECOVERY_NODE_HPP_
 
 #include <string>
 #include "behaviortree_cpp/control_node.h"
 
-namespace nav2_bt_navigator
+namespace nav2_behavior_tree
 {
 /**
  * @brief The RecoveryNode has only two children and returns SUCCESS if and only if the first child
@@ -53,6 +53,6 @@ private:
   BT::NodeStatus tick() override;
   void halt() override;
 };
-}  // namespace nav2_bt_navigator
+}  // namespace nav2_behavior_tree
 
-#endif  // NAV2_BT_NAVIGATOR__RECOVERY_NODE_HPP_
+#endif  // NAV2_BEHAVIOR_TREE__RECOVERY_NODE_HPP_
